+++ CHANGELOG.md
@@ -9,6 +9,8 @@ to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 ### Added
 - All delegations list by address
+- All delegators list by address
+- Delegation summary by address
 
 ## [0.1.14] - 2024-10-08